Create RED-Win64-Shipping.CT
This is a small table to allow the user to lock the controller allocator function that forces PvE on offline local lobby. Just enable the script that replaces it with a NOP and CPU vs CPU will be possible. The only issue is that I was unable to find the pointer to the results so every match will be a defeat.